Create an API for testutils directory
